# documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Node js test api",
        "description": "Ventaja Government Pay Test apis node js",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "contact": {
            "name": "appscrip",
            "email": "jagannath@appscrip.co"
        }
    },
    "components": {
        "securitySchemes": [
            {
                "bearerAuth": {
                    "type": "http",
                    "description": "access token",
                    "scheme": "bearer",
                    "bearerFormat": "JWT",
                    "in": "header"
                }
            }
        ]
    },
    "security": {},
    "servers": [
        "http://localhost:8000/"
    ],
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/ibg/validate-short-term-loan": {
            "post": {
                "tags": [
                    "Ibg-Payment"
                ],
                "summary": "Validate short term loan",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "memberId": {
                                    "type": "string",
                                    "example": "121074345129"
                                },
                                "lastName": {
                                    "type": "string",
                                    "example": "ROCHE"
                                },
                                "firstName": {
                                    "type": "string",
                                    "example": "EMILY"
                                },
                                "middleName": {
                                    "type": "string",
                                    "example": "ARAGON"
                                },
                                "birthDate": {
                                    "type": "string",
                                    "example": "1978-01-01"
                                },
                                "mobileNumber": {
                                    "type": "string",
                                    "example": ""
                                },
                                "emailAddress": {
                                    "type": "string",
                                    "example": "juan.delacruz@email.com"
                                },
                                "memberType": {
                                    "type": "number",
                                    "example": 0
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 100.0
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "transactionNumber": {
                                    "example": "string"
                                },
                                "referenceNumber": {
                                    "example": "string"
                                },
                                "dateEntry": {
                                    "example": "yyyy-MM-ddTHH:mm:ss"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/ibg/process-short-term-loan": {
            "post": {
                "tags": [
                    "Ibg-Payment"
                ],
                "summary": "Process short term loan",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "memberId": {
                                    "type": "string",
                                    "example": "121074345129"
                                },
                                "lastName": {
                                    "type": "string",
                                    "example": "ROCHE"
                                },
                                "firstName": {
                                    "type": "string",
                                    "example": "EMILY"
                                },
                                "middleName": {
                                    "type": "string",
                                    "example": "ARAGON"
                                },
                                "birthDate": {
                                    "type": "string",
                                    "example": "1978-01-01"
                                },
                                "mobileNumber": {
                                    "type": "string",
                                    "example": ""
                                },
                                "emailAddress": {
                                    "type": "string",
                                    "example": "juan.delacruz@email.com"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 100.0
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "transactionNumber": {
                                    "example": "string"
                                },
                                "referenceNumber": {
                                    "example": "string"
                                },
                                "dateEntry": {
                                    "example": "yyyy-MM-ddTHH:mm:ss"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/ibg/validate-housing-loan": {
            "post": {
                "tags": [
                    "Ibg-Payment"
                ],
                "summary": "Process Housing loan",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "memberId": {
                                    "type": "string",
                                    "example": "121074345129"
                                },
                                "lastName": {
                                    "type": "string",
                                    "example": "ROCHE"
                                },
                                "firstName": {
                                    "type": "string",
                                    "example": "EMILY"
                                },
                                "middleName": {
                                    "type": "string",
                                    "example": "ARAGON"
                                },
                                "birthDate": {
                                    "type": "string",
                                    "example": "1978-01-01"
                                },
                                "mobileNumber": {
                                    "type": "string",
                                    "example": ""
                                },
                                "emailAddress": {
                                    "type": "string",
                                    "example": "juan.delacruz@email.com"
                                },
                                "memberType": {
                                    "type": "number",
                                    "example": 0
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 100.0
                                },
                                "startMonth": {
                                    "type": "number",
                                    "example": 2021
                                },
                                "startYear": {
                                    "type": "number",
                                    "example": 4
                                },
                                "endYear": {
                                    "type": "number",
                                    "example": 2021
                                },
                                "endMonth": {
                                    "type": "number",
                                    "example": 6
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "transactionNumber": {
                                    "example": "string"
                                },
                                "referenceNumber": {
                                    "example": "string"
                                },
                                "dateEntry": {
                                    "example": "yyyy-MM-ddTHH:mm:ss"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/ibg/process-housing-loan": {
            "post": {
                "tags": [
                    "Ibg-Payment"
                ],
                "summary": "Process Housing loan",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "memberId": {
                                    "type": "string",
                                    "example": "121074345129"
                                },
                                "lastName": {
                                    "type": "string",
                                    "example": "ROCHE"
                                },
                                "firstName": {
                                    "type": "string",
                                    "example": "EMILY"
                                },
                                "middleName": {
                                    "type": "string",
                                    "example": "ARAGON"
                                },
                                "birthDate": {
                                    "type": "string",
                                    "example": "1978-01-01"
                                },
                                "mobileNumber": {
                                    "type": "string",
                                    "example": ""
                                },
                                "emailAddress": {
                                    "type": "string",
                                    "example": "juan.delacruz@email.com"
                                },
                                "memberType": {
                                    "type": "number",
                                    "example": 0
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 100.0
                                },
                                "startMonth": {
                                    "type": "number",
                                    "example": 2021
                                },
                                "startYear": {
                                    "type": "number",
                                    "example": 4
                                },
                                "endYear": {
                                    "type": "number",
                                    "example": 2021
                                },
                                "endMonth": {
                                    "type": "number",
                                    "example": 6
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "transactionNumber": {
                                    "example": "string"
                                },
                                "referenceNumber": {
                                    "example": "string"
                                },
                                "dateEntry": {
                                    "example": "yyyy-MM-ddTHH:mm:ss"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "400": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "number",
                    "example": 131
                },
                "message": {
                    "type": "string",
                    "example": "Duplicate transaction"
                },
                "description": {
                    "type": "string",
                    "example": "Applicable date has an overlapping payment"
                },
                "status": {
                    "type": "number",
                    "example": 400
                }
            }
        },
        "500": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error!"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        }
    }
}